requirements.txt
numpy>=1.16.6
tqdm>=4.43.0
demjson==2.2.4
torchtext==0.5.0
GitPython==3.1.0
requests==2.23.0
fasttext==0.9.1
nltk==3.4.5
editdistance==0.5.3
sklearn==0.0
omegaconf==2.0.1rc4
lmdb==0.98
termcolor==1.1.0
